First pass at writing a 'patchy' background subtraction algorithm.
Paul and I had a brief discussion on Friday (Oct 8). The idea here is to find regions laying outside of the machine-learning generated bounding boxes and consider those 'background'. This process can be applied to the last N frames to get an average of what the background is. This averaged background array can then be used as you normally would to perform background image subtraction. Note I haven't tested this yet using the ML algorithm and a set of test RBC images.