Found a bug where after checking if the bounding box coordinates are within bounds (using utils.out_of_bounds()), the bounding box was not actually being updated in the objects being returned.
The objects that are generated and returned by the detection function are namedtuples, which are immutable. Unsure why this was decided on as the data structure to use for these return types. As a temporary fix, I currently create a new object (writing over the existing one) using the updated bounding box.