defensively restart the autofocus model and the cell diagnosis model
IJ found a bug where the start of the predictions handler would have some references to very high image numbers.
Perhaps this is due to the NCSModel thread pool executors not being reset between runs?
This PR defensively resets them, and also shuts them down at the end of an experiment (but that should be unecessary,
since the GC will take care of it anyways)