Update the flow control routine so that the generator doesn't fail (StopIteration exception) when the syringe reaches its maximum position. Instead, return a boolean indicating whether the syringe is able to be moved (if False, this means the syringe a CantReachTargetFlowrate exception was raised further down the stack, meaning the syringe can't move any further and we are below the target flowrate).
This commit changes the behaviour of the flow control routine so that it yields the syringe move boolean. This allows the routine to continue to accept images
and measure flowrate, even if the syringe cannot move. Previously, when the exception was raised, the generator dies. Reinitializing the generator meant restarting the flow control tracking (but
it would inevitably keep dying as the syringe kept raising its CantMove exception).